Update link for 'How To Host Questionable Websites'
Updated links with archived WaybackMachine links since WebOas.is is no longer online. Note that the links contain different versions of the document. The only link displaying a version number is the txt link, as v5.15. However, the txt link is the oldest archived version, so the actual document version is likely higher.